+++ packages/google_maps_flutter/google_maps_flutter_web/lib/src/convert.dart
@@ -98,9 +98,15 @@ gmaps.MapTypeId _gmapTypeIDForPluginType(MapType type) {
       return gmaps.MapTypeId.HYBRID;
     case MapType.normal:
     case MapType.none:
-    default:
       return gmaps.MapTypeId.ROADMAP;
   }
+  // The enum comes from a different package, which could get a new value at
+  // any time, so provide a fallback that ensures this won't break when used
+  // with a version that contains new values. This is deliberately outside
+  // the switch rather than a `default` so that the linter will flag the
+  // switch as needing an update.
+  // ignore: dead_code
+  return gmaps.MapTypeId.ROADMAP;
 }
 
 gmaps.MapOptions _applyInitialPosition(