+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/ScrollListenerFlutterApiImpl.java
@@ -4,7 +4,6 @@
 
 package io.flutter.plugins.webviewflutter;
 
-
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.ScrollListenerFlutterApi;
 
@@ -14,37 +13,31 @@ import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.ScrollListenerF
  * <p>Passes arguments of callbacks methods from a {@link ScrollListener} to Dart.
  */
 public class ScrollListenerFlutterApiImpl extends ScrollListenerFlutterApi {
-    private final InstanceManager instanceManager;
+  private final InstanceManager instanceManager;
 
-    /**
-     * Creates a Flutter api that sends messages to Dart.
-     *
-     * @param binaryMessenger handles sending messages to Dart
-     * @param instanceManager maintains instances stored to communicate with Dart objects
-     */
-    public ScrollListenerFlutterApiImpl(
-            BinaryMessenger binaryMessenger, InstanceManager instanceManager) {
-        super(binaryMessenger);
-        this.instanceManager = instanceManager;
-    }
+  /**
+   * Creates a Flutter api that sends messages to Dart.
+   *
+   * @param binaryMessenger handles sending messages to Dart
+   * @param instanceManager maintains instances stored to communicate with Dart objects
+   */
+  public ScrollListenerFlutterApiImpl(
+      BinaryMessenger binaryMessenger, InstanceManager instanceManager) {
+    super(binaryMessenger);
+    this.instanceManager = instanceManager;
+  }
 
-    /** Passes arguments from {@link ScrollListener#onScrollPosChange} to Dart. */
-    public void onScrollPosChange(
-            ScrollListener scrollChangeListener,
-            long x,
-            long y,
-            Reply<Void> callback) {
-        onScrollPosChange(
-                getIdentifierForListener(scrollChangeListener),
-                x, y,
-                callback);
-    }
+  /** Passes arguments from {@link ScrollListener#onScrollPosChange} to Dart. */
+  public void onScrollPosChange(
+      ScrollListener scrollChangeListener, long x, long y, Reply<Void> callback) {
+    onScrollPosChange(getIdentifierForListener(scrollChangeListener), x, y, callback);
+  }
 
-    private long getIdentifierForListener(ScrollListener listener) {
-        final Long identifier = instanceManager.getIdentifierForStrongReference(listener);
-        if (identifier == null) {
-            throw new IllegalStateException("Could not find identifier for OnScrollChangeListener.");
-        }
-        return identifier;
+  private long getIdentifierForListener(ScrollListener listener) {
+    final Long identifier = instanceManager.getIdentifierForStrongReference(listener);
+    if (identifier == null) {
+      throw new IllegalStateException("Could not find identifier for OnScrollChangeListener.");
     }
+    return identifier;
+  }
 }