+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/WebViewExtendedApi.java
@@ -5,9 +5,7 @@
 package io.flutter.plugins.webviewflutter;
 
 import androidx.annotation.Nullable;
-/**
- * Define extending APIs for {@link android.webkit.WebView}
- */
+/** Define extending APIs for {@link android.webkit.WebView} */
 public interface WebViewExtendedApi {
-    void setScrollListener(@Nullable ScrollListener onScrollChangeListener);
+  void setScrollListener(@Nullable ScrollListener onScrollChangeListener);
 }