+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/WebViewHostApiImpl.java
@@ -12,15 +12,13 @@ import android.webkit.DownloadListener;
 import android.webkit.WebChromeClient;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
-
-import java.util.Map;
-import java.util.Objects;
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.platform.PlatformView;
 import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.WebViewHostApi;
+import java.util.Map;
+import java.util.Objects;
 
 /**
  * Host api implementation for {@link WebView}.
@@ -79,11 +77,11 @@ public class WebViewHostApiImpl implements WebViewHostApi {
   }
 
   /** Implementation of {@link WebView} that can be used as a Flutter {@link PlatformView}s. */
-  public static class WebViewPlatformView extends WebView implements PlatformView, WebViewExtendedApi {
+  public static class WebViewPlatformView extends WebView
+      implements PlatformView, WebViewExtendedApi {
     private WebViewClient currentWebViewClient;
     private WebChromeClientHostApiImpl.SecureWebChromeClient currentWebChromeClient;
-    private @Nullable
-    ScrollListener onScrollChangeListener;
+    private @Nullable ScrollListener onScrollChangeListener;
 
     /**
      * Creates a {@link WebViewPlatformView}.
@@ -157,8 +155,7 @@ public class WebViewHostApiImpl implements WebViewHostApi {
       implements PlatformView, WebViewExtendedApi {
     private WebViewClient currentWebViewClient;
     private WebChromeClientHostApiImpl.SecureWebChromeClient currentWebChromeClient;
-    private @Nullable
-    ScrollListener onScrollChangeListener;
+    private @Nullable ScrollListener onScrollChangeListener;
 
     /**
      * Creates a {@link InputAwareWebViewPlatformView}.
@@ -455,9 +452,11 @@ public class WebViewHostApiImpl implements WebViewHostApi {
   @Override
   public void setScrollListener(@NonNull Long instanceId, @Nullable Long scrollListenerInstanceId) {
     final WebView webView = (WebView) instanceManager.getInstance(instanceId);
-    if(webView instanceof WebViewExtendedApi) {
-      if(scrollListenerInstanceId != null) {
-        ((WebViewExtendedApi) webView).setScrollListener((ScrollListener) instanceManager.getInstance(scrollListenerInstanceId));
+    if (webView instanceof WebViewExtendedApi) {
+      if (scrollListenerInstanceId != null) {
+        ((WebViewExtendedApi) webView)
+            .setScrollListener(
+                (ScrollListener) instanceManager.getInstance(scrollListenerInstanceId));
       } else {
         ((WebViewExtendedApi) webView).setScrollListener(null);
       }