+++ packages/webview_flutter/webview_flutter_android/android/src/test/java/io/flutter/plugins/webviewflutter/ScrollListenerTest.java
@@ -4,6 +4,12 @@
 
 package io.flutter.plugins.webviewflutter;
 
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.verify;
+
+import io.flutter.plugins.webviewflutter.ScrollListenerHostApiImpl.ScrollListenerCreator;
+import io.flutter.plugins.webviewflutter.ScrollListenerHostApiImpl.ScrollListenerImpl;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
@@ -11,13 +17,6 @@ import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
-
-import io.flutter.plugins.webviewflutter.ScrollListenerHostApiImpl.ScrollListenerImpl;
-import io.flutter.plugins.webviewflutter.ScrollListenerHostApiImpl.ScrollListenerCreator;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.verify;
 
 public class ScrollListenerTest {
   @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
@@ -35,8 +34,7 @@ public class ScrollListenerTest {
     final ScrollListenerCreator ScrollListenerCreator =
         new ScrollListenerCreator() {
           @Override
-          public ScrollListenerImpl createScrollListener(
-              ScrollListenerFlutterApiImpl flutterApi) {
+          public ScrollListenerImpl createScrollListener(ScrollListenerFlutterApiImpl flutterApi) {
             scrollListener = super.createScrollListener(flutterApi);
             return scrollListener;
           }
@@ -55,11 +53,6 @@ public class ScrollListenerTest {
   @Test
   public void scrollPosChange() {
     scrollListener.onScrollPosChange(1, 2);
-    verify(mockFlutterApi).onScrollPosChange(
-            eq(scrollListener),
-            eq(1L),
-            eq(2L),
-            any()
-    );
+    verify(mockFlutterApi).onScrollPosChange(eq(scrollListener), eq(1L), eq(2L), any());
   }
 }