+++ packages/webview_flutter/webview_flutter_android/example/lib/main.dart
@@ -32,7 +32,6 @@ The navigation delegate is set to block navigation to the youtube website.
 </body>
 </html>
 ''';
-
 const String kLocalExamplePage = '''
 <!DOCTYPE html>
 <html lang="en">
@@ -51,7 +50,6 @@ const String kLocalExamplePage = '''
 </body>
 </html>
 ''';
-
 const String kTransparentBackgroundPage = '''
 <!DOCTYPE html>
 <html>
@@ -75,20 +73,16 @@ const String kTransparentBackgroundPage = '''
 
 class WebViewExample extends StatefulWidget {
   const WebViewExample({Key? key, this.cookieManager}) : super(key: key);
-
   final PlatformWebViewCookieManager? cookieManager;
-
   @override
   State<WebViewExample> createState() => _WebViewExampleState();
 }
 
 class _WebViewExampleState extends State<WebViewExample> {
   late final PlatformWebViewController _controller;
-
   @override
   void initState() {
     super.initState();
-
     _controller = PlatformWebViewController(
       AndroidWebViewControllerCreationParams(),
     )
@@ -146,8 +140,8 @@ Page resource error:
     return Scaffold(
       backgroundColor: const Color(0xFF4CAF50),
       appBar: AppBar(
-        title: const Text('Flutter WebView example'),
-        // This drop down menu demonstrates that Flutter widgets can be shown over the web view.
+        title: const Text(
+            'Flutter WebView example'), // This drop down menu demonstrates that Flutter widgets can be shown over the web view.
         actions: <Widget>[
           NavigationControls(webViewController: _controller),
           SampleMenu(
@@ -167,9 +161,11 @@ Page resource error:
     return FloatingActionButton(
       onPressed: () async {
         final String? url = await _controller.currentUrl();
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text('Favorited $url')),
-        );
+        if (context.mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(content: Text('Favorited $url')),
+          );
+        }
       },
       child: const Icon(Icons.favorite),
     );
@@ -202,10 +198,8 @@ class SampleMenu extends StatelessWidget {
               const PlatformWebViewCookieManagerCreationParams(),
             ),
         super(key: key);
-
   final PlatformWebViewController webViewController;
   late final PlatformWebViewCookieManager cookieManager;
-
   @override
   Widget build(BuildContext context) {
     return PopupMenuButton<MenuOptions>(
@@ -322,30 +316,34 @@ class SampleMenu extends StatelessWidget {
   Future<void> _onListCookies(BuildContext context) async {
     final String cookies = await webViewController
         .runJavaScriptReturningResult('document.cookie') as String;
-    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-      content: Column(
-        mainAxisAlignment: MainAxisAlignment.end,
-        mainAxisSize: MainAxisSize.min,
-        children: <Widget>[
-          const Text('Cookies:'),
-          _getCookieList(cookies),
-        ],
-      ),
-    ));
+    if (context.mounted) {
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+        content: Column(
+          mainAxisAlignment: MainAxisAlignment.end,
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            const Text('Cookies:'),
+            _getCookieList(cookies),
+          ],
+        ),
+      ));
+    }
   }
 
   Future<void> _onAddToCache(BuildContext context) async {
     await webViewController.runJavaScript(
       'caches.open("test_caches_entry"); localStorage["test_localStorage"] = "dummy_entry";',
     );
-    ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-      content: Text('Added a test entry to cache.'),
-    ));
+    if (context.mounted) {
+      ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
+        content: Text('Added a test entry to cache.'),
+      ));
+    }
   }
 
   Future<void> _onListCache() {
-    return webViewController.runJavaScript('caches.keys()'
-        // ignore: missing_whitespace_between_adjacent_strings
+    return webViewController.runJavaScript(
+        'caches.keys()' // ignore: missing_whitespace_between_adjacent_strings
         '.then((cacheKeys) => JSON.stringify({"cacheKeys" : cacheKeys, "localStorage" : localStorage}))'
         '.then((caches) => Toaster.postMessage(caches))');
   }
@@ -353,9 +351,11 @@ class SampleMenu extends StatelessWidget {
   Future<void> _onClearCache(BuildContext context) async {
     await webViewController.clearCache();
     await webViewController.clearLocalStorage();
-    ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
-      content: Text('Cache cleared.'),
-    ));
+    if (context.mounted) {
+      ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
+        content: Text('Cache cleared.'),
+      ));
+    }
   }
 
   Future<void> _onClearCookies(BuildContext context) async {
@@ -364,9 +364,11 @@ class SampleMenu extends StatelessWidget {
     if (!hadCookies) {
       message = 'There are no cookies.';
     }
-    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-      content: Text(message),
-    ));
+    if (context.mounted) {
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+        content: Text(message),
+      ));
+    }
   }
 
   Future<void> _onNavigationDelegateExample() {
@@ -441,10 +443,8 @@ class SampleMenu extends StatelessWidget {
     final String tmpDir = (await getTemporaryDirectory()).path;
     final File indexFile = File(
         <String>{tmpDir, 'www', 'index.html'}.join(Platform.pathSeparator));
-
     await indexFile.create(recursive: true);
     await indexFile.writeAsString(kLocalExamplePage);
-
     return indexFile.path;
   }
 }
@@ -452,9 +452,7 @@ class SampleMenu extends StatelessWidget {
 class NavigationControls extends StatelessWidget {
   const NavigationControls({Key? key, required this.webViewController})
       : super(key: key);
-
   final PlatformWebViewController webViewController;
-
   @override
   Widget build(BuildContext context) {
     return Row(
@@ -465,10 +463,11 @@ class NavigationControls extends StatelessWidget {
             if (await webViewController.canGoBack()) {
               await webViewController.goBack();
             } else {
-              ScaffoldMessenger.of(context).showSnackBar(
-                const SnackBar(content: Text('No back history item')),
-              );
-              return;
+              if (context.mounted) {
+                ScaffoldMessenger.of(context).showSnackBar(
+                  const SnackBar(content: Text('No back history item')),
+                );
+              }
             }
           },
         ),
@@ -478,10 +477,11 @@ class NavigationControls extends StatelessWidget {
             if (await webViewController.canGoForward()) {
               await webViewController.goForward();
             } else {
-              ScaffoldMessenger.of(context).showSnackBar(
-                const SnackBar(content: Text('No forward history item')),
-              );
-              return;
+              if (context.mounted) {
+                ScaffoldMessenger.of(context).showSnackBar(
+                  const SnackBar(content: Text('No forward history item')),
+                );
+              }
             }
           },
         ),