+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview.dart
@@ -20,6 +20,8 @@ import 'android_webview.pigeon.dart';
 import 'android_webview_api_impls.dart';
 import 'instance_manager.dart';
 
+export 'android_webview_api_impls.dart' show FileChooserMode;
+
 /// Root of the Java class hierarchy.
 ///
 /// See https://docs.oracle.com/javase/8/docs/api/java/lang/Object.html.
@@ -397,7 +399,7 @@ class WebView extends JavaObject {
 
   /// Sets the scroll listener to this WebView.
   Future<void> setScrollListener(ScrollListener? scrollListener) {
-    if(scrollListener != null) {
+    if (scrollListener != null) {
       ScrollListener.api.createFromInstance(scrollListener);
     }
     return api.setScrollListenerFromInstance(this, scrollListener);
@@ -907,7 +909,8 @@ class ScrollListener extends JavaObject {
 /// Handles JavaScript dialogs, favicons, titles, and the progress for [WebView].
 class WebChromeClient extends JavaObject {
   /// Constructs a [WebChromeClient].
-  WebChromeClient({this.onProgressChanged}) : super.detached() {
+  WebChromeClient({this.onProgressChanged, this.onShowFileChooser})
+      : super.detached() {
     AndroidWebViewFlutterApis.instance.ensureSetUp();
     api.createFromInstance(this);
   }
@@ -917,7 +920,10 @@ class WebChromeClient extends JavaObject {
   ///
   /// This should only be used by subclasses created by this library or to
   /// create copies.
-  WebChromeClient.detached({this.onProgressChanged}) : super.detached();
+  WebChromeClient.detached({
+    this.onProgressChanged,
+    this.onShowFileChooser,
+  }) : super.detached();
 
   /// Pigeon Host Api implementation for [WebChromeClient].
   @visibleForTesting
@@ -926,9 +932,95 @@ class WebChromeClient extends JavaObject {
   /// Notify the host application that a file should be downloaded.
   final void Function(WebView webView, int progress)? onProgressChanged;
 
+  /// Indicates the client should show a file chooser.
+  ///
+  /// To handle the request for a file chooser with this callback, passing true
+  /// to [setSynchronousReturnValueForOnShowFileChooser] is required. Otherwise,
+  /// the returned list of strings will be ignored and the client will use the
+  /// default handling of a file chooser request.
+  ///
+  /// Only invoked on Android versions 21+.
+  final Future<List<String>> Function(
+    WebView webView,
+    FileChooserParams params,
+  )? onShowFileChooser;
+
+  /// Sets the required synchronous return value for the Java method,
+  /// `WebChromeClient.onShowFileChooser(...)`.
+  ///
+  /// The Java method, `WebChromeClient.onShowFileChooser(...)`, requires
+  /// a boolean to be returned and this method sets the returned value for all
+  /// calls to the Java method.
+  ///
+  /// Setting this to true indicates that all file chooser requests should be
+  /// handled by [onShowFileChooser] and the returned list of Strings will be
+  /// returned to the WebView. Otherwise, the client will use the default
+  /// handling and the returned value in [onShowFileChooser] will be ignored.
+  ///
+  /// Requires [onShowFileChooser] to be nonnull.
+  ///
+  /// Defaults to false.
+  Future<void> setSynchronousReturnValueForOnShowFileChooser(
+    bool value,
+  ) {
+    if (value && onShowFileChooser == null) {
+      throw StateError(
+        'Setting this to true requires `onShowFileChooser` to be nonnull.',
+      );
+    }
+    return api.setSynchronousReturnValueForOnShowFileChooserFromInstance(
+      this,
+      value,
+    );
+  }
+
   @override
   WebChromeClient copy() {
-    return WebChromeClient.detached(onProgressChanged: onProgressChanged);
+    return WebChromeClient.detached(
+      onProgressChanged: onProgressChanged,
+      onShowFileChooser: onShowFileChooser,
+    );
+  }
+}
+
+/// Parameters received when a [WebChromeClient] should show a file chooser.
+///
+/// See https://developer.android.com/reference/android/webkit/WebChromeClient.FileChooserParams.
+class FileChooserParams extends JavaObject {
+  /// Constructs a [FileChooserParams] without creating the associated Java
+  /// object.
+  ///
+  /// This should only be used by subclasses created by this library or to
+  /// create copies.
+  FileChooserParams.detached({
+    required this.isCaptureEnabled,
+    required this.acceptTypes,
+    required this.filenameHint,
+    required this.mode,
+    super.binaryMessenger,
+    super.instanceManager,
+  }) : super.detached();
+
+  /// Preference for a live media captured value (e.g. Camera, Microphone).
+  final bool isCaptureEnabled;
+
+  /// A list of acceptable MIME types.
+  final List<String> acceptTypes;
+
+  /// The file name of a default selection if specified, or null.
+  final String? filenameHint;
+
+  /// Mode of how to select files for a file chooser.
+  final FileChooserMode mode;
+
+  @override
+  FileChooserParams copy() {
+    return FileChooserParams.detached(
+      isCaptureEnabled: isCaptureEnabled,
+      acceptTypes: acceptTypes,
+      filenameHint: filenameHint,
+      mode: mode,
+    );
   }
 }