+++ packages/webview_flutter/webview_flutter_android/test/android_webview_test.mocks.dart
@@ -315,6 +315,21 @@ class MockTestWebChromeClientHostApi extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+  @override
+  void setSynchronousReturnValueForOnShowFileChooser(
+    int? instanceId,
+    bool? value,
+  ) =>
+      super.noSuchMethod(
+        Invocation.method(
+          #setSynchronousReturnValueForOnShowFileChooser,
+          [
+            instanceId,
+            value,
+          ],
+        ),
+        returnValueForMissingStub: null,
+      );
 }
 
 /// A class which mocks [TestWebSettingsHostApi].
@@ -1021,6 +1036,16 @@ class MockWebChromeClient extends _i1.Mock implements _i2.WebChromeClient {
     _i1.throwOnMissingStub(this);
   }
 
+  @override
+  _i5.Future<void> setSynchronousReturnValueForOnShowFileChooser(bool? value) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #setSynchronousReturnValueForOnShowFileChooser,
+          [value],
+        ),
+        returnValue: _i5.Future<void>.value(),
+        returnValueForMissingStub: _i5.Future<void>.value(),
+      ) as _i5.Future<void>);
   @override
   _i2.WebChromeClient copy() => (super.noSuchMethod(
         Invocation.method(