+++ packages/camera/camera/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.10.3
+
+* Adds back use of Optional type.
+
+## 0.10.2+1
+
+* Updates code for stricter lint checks.
+
+## 0.10.2
+
+* Implements option to also stream when recording a video.
+
 ## 0.10.1
 
 * Remove usage of deprecated quiver Optional type.
@@ -27,7 +39,7 @@
 ## 0.10.0
 
 * **Breaking Change** Bumps default camera_web package version, which updates permission exception code from `cameraPermission` to `CameraAccessDenied`.
-* **Breaking Change** Bumps default camera_android package version, which updates permission exception code from `cameraPermission` to 
+* **Breaking Change** Bumps default camera_android package version, which updates permission exception code from `cameraPermission` to
   `CameraAccessDenied` and `AudioAccessDenied`.
 * Ignores unnecessary import warnings in preparation for [upcoming Flutter changes](https://github.com/flutter/flutter/pull/106316).